Resolve build warnings

+++ src/backend/TrafficCourts/Test/Citizen.Service/Mappings/NoticeOfDisputeToMessageContractMappingTests.cs
@@ -72,7 +72,7 @@ namespace TrafficCourts.Test.Citizen.Service.Mappings
 
             foreach (var count in actual.DisputeCounts)
             {
-                Assert.Contains(noticeOfDispute.DisputeCounts!, _ => (int)_.PleaCode == (int)count.PleaCode);
+                Assert.Contains(noticeOfDispute.DisputeCounts!, _ => _.PleaCode == count.PleaCode);
                 Assert.Contains(noticeOfDispute.DisputeCounts!, _ => _.CountNo == count.CountNo);
                 Assert.Contains(noticeOfDispute.DisputeCounts!, _ => _.RequestCourtAppearance == count.RequestCourtAppearance);
                 Assert.Contains(noticeOfDispute.DisputeCounts!, _ => _.RequestReduction == count.RequestReduction);
